add deflate-stream and improve inflate-stream
The deflate stream is in interface to the libz compression library
found on most unix systems and available on Windows system.
In this commit we've only tested on Linux

Are there user visible changes in this commit?  yes

Is bootstrapping needed?   no

Are tests included for new features? no, but aserve's tests test this

Tests run:  none

<release-note>
new deflate-stream and improved inflate-stream

The deflate-stream is documented in deflate.txt in this module.

The inflate stream change is that there is a new argument
to make-instance of an inflate-stream: :skip-gzip-header
with description:
Streams to be inflated often begin with a gzip header which
must be skipped over before inflation can begin.  Now you can
pass the argument :skip-gzip-header t to the call to
make-instance of an inflate-stream.  If the stream begins with
a gzip header then that header will be skipped.  If the stream
doesn't begin with a gzip header then this argument has no effect

</release-note>

<documentation>
see deflate.txt for documentation on the deflate-stream
</documentation>

Change-Id: I37fdea747a8e31a414fc7c986a38190de562181a