push :zlib-deflate on *features* if deflate.fasl is loaded
In order for code needing compression to check if the deflate.fasl file was
loaded successfully deflate.fasl puts :zlib-deflate on the *features* list.
We use zlib rather than libz since the authors refer to this code
as zlib and I suppose they didn't want to name the library on
Linux libzlib so it was shortened to libz.  It may have other names
on other operating system.

Are there user visible changes in this commit?  yes

Is bootstrapping needed?   no

Are tests included for new features?  no (a trival change)

Tests run:  none

<release-note>
add :zlib-deflate to *features*

If deflate.fasl was sucessfully loaded, which means that the zlib
foreign library (called libz.so on Linux) was found and loaded as well,
then :zlib-deflate will be found on *features*

</release-note>

Change-Id: Iee9be53bb07b7581ca579fdc84ece26e73683725